config: many more tests and coverage fixes
